添加图形验证码

+++ PolyFinancial-TXJ-admin-html/js/ctrl/backStageCtrl.js
@@ -1,6 +1,8 @@
-app.controller("backStageCtrl",function ($http,$state,yahaha) {
+app.controller("backStageCtrl",function ($http,$state,serviceHTTP) {
 
     var vm = this;
+
+    vm.name = sessionStorage.getItem("userName");
     // serviceHTTP.moduleHTTP().then(function successCallback(res) {
     //     // 请求成功执行代码
     //     console.log(response.data.message);
@@ -18,30 +20,30 @@ app.controller("backStageCtrl",function ($http,$state,yahaha) {
         $state.go('login');
     }
     // 服务器返回的模块数据
-    // vm.sideBar = JSON.parse(sessionStorage.getItem("sideBar"))
-    // console.log(vm.sideBar);
-
-    vm.yahaha= yahaha;
-    vm.sideBar= [];//用于存储处理过后的数组对象
-    console.log(vm.yahaha);
-    function change(e){
-        for(var i = 0 ;i<e.length;i++){//遍历每一个元素
-            if(e[i].parentID === 0){//当为父级元素时开始判断
-                var obj = {};//创建一个对象用于存储父级标签名，子级标签的信息
-                obj.sideBarTitle = e[i].name;
-                obj.id = e[i].id;
-                obj.sideBarContent = [];//为对象添加属性
-                for(var j = 0;j<e.length;j++){//开始第二次遍历，
-                    if(e[j].parentID === e[i].id){//当元素的parentd和  父级元素id一致时
-                        obj.sideBarContent.push(e[j])//把这个元素添加到对象的一条属性中去，用于第二次repeat
-                    }
-                }
-                vm.sideBar.push(obj);//把obj添加到sideBar中去。
-            }
-        }
-    }
-    change(vm.yahaha);
+    vm.sideBar = JSON.parse(sessionStorage.getItem("sideBar"));
     console.log(vm.sideBar);
+
+    // vm.yahaha= yahaha;
+    // vm.sideBar= [];//用于存储处理过后的数组对象
+    // console.log(vm.yahaha);
+    // function change(e){
+    //     for(var i = 0 ;i<e.length;i++){//遍历每一个元素
+    //         if(e[i].parentID === 0){//当为父级元素时开始判断
+    //             var obj = {};//创建一个对象用于存储父级标签名，子级标签的信息
+    //             obj.sideBarTitle = e[i].name;
+    //             obj.id = e[i].id;
+    //             obj.sideBarContent = [];//为对象添加属性
+    //             for(var j = 0;j<e.length;j++){//开始第二次遍历，
+    //                 if(e[j].parentID === e[i].id){//当元素的parentd和  父级元素id一致时
+    //                     obj.sideBarContent.push(e[j])//把这个元素添加到对象的一条属性中去，用于第二次repeat
+    //                 }
+    //             }
+    //             vm.sideBar.push(obj);//把obj添加到sideBar中去。
+    //         }
+    //     }
+    // }
+    // change(vm.yahaha);
+    // console.log(vm.sideBar);
 
     vm.titleIndex = function (e) {
         vm.sideBarTitleIndex = (vm.sideBarTitleIndex === e)? undefined : e;
@@ -59,9 +61,21 @@ app.controller("backStageCtrl",function ($http,$state,yahaha) {
 
     vm.exit = function(){
         if(confirm("确定要退出登录嘛")){
-            alert("ヾ(￣▽￣)Bye~Bye~");
-            sessionStorage.clear();
-            $state.go('login');
+            serviceHTTP.logoutHTTP().then(function successCallback(response) {
+                // 请求成功执行代码
+                console.log(response);
+                if(response.data.message === "success") {
+                    alert("ヾ(￣▽￣)Bye~Bye~");
+                    sessionStorage.clear();
+                    $state.go('login');
+                }
+                else {
+                    alert("服务器歇菜了呢，请稍后重试")
+                }
+            }, function errorCallback(res) {
+                // 请求失败执行代码
+            });
+
         }
 
     }
